requirements.txt
fastapi
uvicorn
yt-dlp
ffmpeg-python
colorama
python-dotenv
pydantic
aiohttp